==> 0. Clean Data.ipynb <==
DATA_PATH = './data/'
LIGHTCURVES_PATH = DATA_PATH + 'lightcurves/'
FEATURES_PATH = DATA_PATH + 'features/'

import numpy as np
import pandas as pd
import measurements, extract
import matplotlib.pyplot as plt

def deleteCopies(df_lcs, copies_dict):
    for original_id, current_copies_list in copies_dict.items():
        df_lcs = df_lcs.drop(current_copies_list, level='ID')
    return df_lcs

def areEqualLightCurves(df_lcs, id1, id2):
    lc1 = df_lcs.loc[id1].sort_values(by='MJD')
    lc1 = lc1[['FLUX', 'FLUX_ERROR', 'MJD']].reset_index(drop=True)
    lc2 = df_lcs.loc[id2].sort_values(by='MJD')
    lc2 = lc2[['FLUX', 'FLUX_ERROR', 'MJD']].reset_index(drop=True)
    return lc1.equals(lc2)

def possiblyRepeatedObservations(df_lcs, min_obs = 1):
    
    df_counts_by_observation = df_lcs.groupby(['FLUX', 'FLUX_ERROR', 'MJD'], as_index=False).size().reset_index(name="ObsCount")
    df_counts_by_observation = df_counts_by_observation[df_counts_by_observation.ObsCount > min_obs]
    return df_counts_by_observation

def possiblyRepeatedByGroups(df_lcs):
    groups = []
    existing = dict()
    df_poss_repeat_obs = possiblyRepeatedObservations(df_lcs)
    for _, row in df_poss_repeat_obs.iterrows():
        current_ids = df_lcs[
            (df_lcs.FLUX == row.FLUX)&(df_lcs.FLUX_ERROR == row.FLUX_ERROR) &(df_lcs.MJD == row.MJD)
        ].index.get_level_values('ID').unique().format()
        if(len(current_ids) > 1) and tuple(current_ids) not in existing:
            groups.append(current_ids)
            existing[tuple(current_ids)] = True
    return groups

def repeatedLightCurves(df_lcs, possibly_by_groups_list):
    copies = {}
    len_possibly_by_groups_list = len(possibly_by_groups_list)
    for i_group, group in enumerate(possibly_by_groups_list):
        if i_group % int(len_possibly_by_groups_list/10) == 0:
            print(i_group, '/', len_possibly_by_groups_list)
        found_as_copy = []
        for i, id1 in enumerate(group):
#             print(i, len(group))
            if id1 not in found_as_copy:
                for j, id2 in enumerate(group[i+1:]):
                    are_equal = areEqualLightCurves(df_lcs, id1, id2)
                    if are_equal and (id1 not in copies or id2 not in copies[id1]): 
                        found_as_copy.append(id2)
                        print(id1, id2)
                        if id1 not in copies: 
                            copies[id1] = []
                        copies[id1].append(id2)
    return copies

### Non-SN

Import non-sn light curves

filename = 'not_sns.pickle'
indir = LIGHTCURVES_PATH; filepath = indir + filename
nsns = pd.read_pickle(filepath)
nsns.shape

nsns.head()

Drop irrelevant columns

nsns = nsns.drop(['classification','zpsys','ra','dec','magnitude','magnitude_error','zp'],axis=1)
nsns.head()

bandpass_elems = ['i','g','r','z']
for elem in bandpass_elems:
    nsns.loc[nsns.bandpass=='sdss'+elem, 'bandpass'] = elem

nsns.groupby(['bandpass']).groups

nsns.head()

nsns.columns = ['MJD','FLUX','FLUX_ERROR','BANDPASS',"ID"]
nsns.shape

Delete rows of blended observations

nsns = nsns.drop_duplicates(['ID','MJD','BANDPASS'], keep='first')
nsns.shape

Replace index into --> [ID(Object), observation_id]

nsns.index.name = 'observation_id'
nsns = nsns.set_index(['ID'], append=True)
nsns = nsns.reorder_levels(['ID', 'observation_id'])
nsns.shape

nsns.head()

Find duplicated light-curves in non-transient dataframe

# Find list containing groups of possibly repeated light curves
# Takes a while...
#groups arent repeated in our case.
possibly_by_groups_list_nt = possiblyRepeatedByGroups(nsns)

copies_nt = repeatedLightCurves(nsns, possibly_by_groups_list_nt)

copies_nt, nsns.shape

dt_nt_lcs = deleteCopies(df_nt_lcs, copies_nt)

df_nt_lcs.shape

nsns.head()

######  TO PICKLE

filename = 'not_sns_clean.pickle'
outdir = LIGHTCURVES_PATH; filepath = outdir + filename
nsns.to_pickle(filepath)

### imulated supernova curves

Import and filter transients

filename = 'sns.pickle'
indir = LIGHTCURVES_PATH; filepath = indir + filename
sns = pd.read_pickle(filepath)
sns.shape

sns.head()

Delete rows of blended observations

# Delete rows of blended observations
sns = sns.drop_duplicates(['ID','MJD','FLT'], keep='first')
sns.shape

Drop irrelevant columns and rename columns to standard 

sns = sns.drop(['FIELD','PHOTFLAG','ZPT','PSF','SIM_MAGOBS'],axis=1)

sns.columns = ['MJD','BANDPASS','FLUX','FLUX_ERROR',"ID"]
sns.head()

Reset Index

sns.index.name = 'observation_id'
sns = sns.rename(columns={'TransientID':'ID'})
sns = sns.set_index(['ID'], append=True)
sns = sns.reorder_levels(['ID', 'observation_id'])
sns.head()

num_objects = sns.index.get_level_values('ID').unique()
len(num_objects)

Find copies

# Takes a while...
possibly_by_groups_list_t = possiblyRepeatedByGroups(sns)

copies_t = repeatedLightCurves(sns, possibly_by_groups_list_t)

len(copies_t.items())

Delete copies

sns= deleteCopies(sns, copies_t)

Show results

num_objects = sns.index.get_level_values('ID').unique()
len(num_objects)

sns.shape

sns.head()

###### TO PICKLE

filename = 'sns_clean.pickle'
outdir = LIGHTCURVES_PATH; filepath = outdir + filename
sns.to_pickle(filepath)

Test repeated have been cleared

possibly_by_groups_list_t_updated = possiblyRepeatedByGroups(sns)

len(possibly_by_groups_list_t_updated)

copies_t = repeatedLightCurves(sns, possibly_by_groups_list_t_updated)

### Measured Supernovas

filename = 'measured_sns.pickle'
indir = LIGHTCURVES_PATH; filepath = indir + filename
msns = pd.read_pickle(filepath)
msns.shape

msns = msns.drop(['classification','zpsys','ra','dec','magnitude','magnitude_error','zp'],axis=1)
msns.head()

msns["ID"] = msns["ID"].apply(lambda x: "MSN_"+str(x))
# df['Date'] = df['Date'].apply(lambda x: int(str(x)[-4:]))

bandpass_elems = ['i','g','r','z']
for elem in bandpass_elems:
    msns.loc[msns.bandpass=='sdss'+elem, 'bandpass'] = elem

msns.groupby(['bandpass']).groups

msns.columns = ['MJD','FLUX','FLUX_ERROR','BANDPASS',"ID"]
msns.shape

msns = msns.drop_duplicates(['ID','MJD','BANDPASS'], keep='first')
msns.shape

msns.index.name = 'observation_id'
msns = msns.set_index(['ID'], append=True)
msns = msns.reorder_levels(['ID', 'observation_id'])
msns.shape

msns.head()

filename = 'measured_sns_clean.pickle'
outdir = LIGHTCURVES_PATH; filepath = outdir + filename
msns.to_pickle(filepath)
