==> 1. Filtering & Feature Extraction.ipynb <==
%reload_ext autoreload
%autoreload 2

DATA_PATH = './data/'
LIGHTCURVES_PATH = DATA_PATH + 'lightcurves/'
FEATURES_PATH = DATA_PATH + 'features/'

import numpy as np
import pandas as pd
import measurements, extract
import matplotlib.pyplot as plt
import inputs2
from log_progress import log_progress

def unique_ids_list(df_lcs):
    return df_lcs.index.get_level_values('ID').unique().format()

def print_num_ids_shape(df_lcs):
    unique_ids = unique_ids_list(df_lcs)
    print('Num IDs: {}  Shape: {}'.format(len(unique_ids), df_lcs.shape))

#### Import

Import __supernova__ lightcurves

filename = 'sns_clean.pickle'
indir = LIGHTCURVES_PATH; filepath = indir + filename
sns = pd.read_pickle(filepath)
print_num_ids_shape(sns)

(sns.index.get_level_values("ID").unique())

Import __non-supernovas__ light curves

filename = 'not_sns_clean.pickle'
indir = LIGHTCURVES_PATH; filepath = indir + filename
nsns = pd.read_pickle(filepath)
print_num_ids_shape(nsns)

Import __measured supernova___ lightcurves

filename = 'measured_sns_clean.pickle'
indir = LIGHTCURVES_PATH; filepath = indir + filename
msns = pd.read_pickle(filepath)
print_num_ids_shape(msns)

msns.head()


filename = "diego_t_clean.pickle"
indir = LIGHTCURVES_PATH; filepath = indir + filename
nt = pd.read_pickle(filepath)
print_num_ids_shape(nt)

nt.tail()

### Add class and set BANDPASS as index


__SN__

sns['class'] = 'sn'
msns['class'] = 'sn'

__Non-SN__

nsns['class'] = 'non-sn'

sns = sns.set_index(['BANDPASS'], append=True)

nsns = nsns.set_index(['BANDPASS'], append=True)

msns= msns.set_index(['BANDPASS'],append=True)

msns.head()

#### Filter

sns.index.get_values

sns.head()

# Make columns index again and then run the filter

def filter_light_curves(sns, min_obs):
#     df_count = df_lcs.groupby(['ID','BANDPASS'], as_index=True).count()
#     df_count['ObsCount'] = df_count['FLUX']
#     df_count = df_count[['ObsCount']]
# #     df_count.head()
#     df_lcs_with_counts = df_lcs.join(df_count, how='inner')
#     # Remove objects with less than min_obs
#     df_filtered = df_lcs_with_counts[df_lcs_with_counts.ObsCount >= min_obs]
    
    df_count = sns.groupby(['ID','BANDPASS'], as_index=True).count()
    df_count['ObsCount'] = df_count['FLUX']
    df_count = df_count[['ObsCount']]

    sns = sns.reset_index(level=['ID', 'BANDPASS'])
    df_count = df_count.reset_index(level=['ID', 'BANDPASS'])

    df_lcs_with_counts = sns.merge(df_count,on=['ID','BANDPASS'])

    df_lcs_with_counts.head()

    df_lcs_with_counts.index.name = 'observation_id'
    df_lcs_with_counts = df_lcs_with_counts.set_index(['ID'], append=True)
    df_lcs_with_counts = df_lcs_with_counts.set_index(['BANDPASS'], append=True)
    df_lcs_with_counts = df_lcs_with_counts.reorder_levels(['ID', 'observation_id','BANDPASS'])
#     # Remove ObsCount
    df_filtered = df_lcs_with_counts[df_lcs_with_counts.ObsCount >= min_obs]
    df_filtered = df_filtered.drop(['ObsCount'], axis=1)
# 

    return df_filtered

def merge_multi(self, df, on):
    return self.reset_index().join(df,on=on,how=inner).set_index(self.index.names)

def sample(df_lcs, num_samples):
    # Set random seed
    np.random.seed(42)
    # Sample non-transient subset of same size as transients
    IDs = np.random.choice(unique_ids_list(df_lcs), size=num_samples, replace=False)
#     print(IDs); return
    df_sampled = df_nont.loc[IDs]
    return df_sampled

Filter __supernova__ light curves

df_sn_5 = filter_light_curves(sns, 5)
print_num_ids_shape(df_sn_5)

del sns

df_sn_5.head()

Filter __non-supernova__ lightcurves

df_nsns_5 = filter_light_curves(nsns, 5)
print_num_ids_shape(df_nsns_5)

del nsns

df_msns_5 = filter_light_curves(msns,5)
print_num_ids_shape(df_msns_5)
del msns

#### Oversample - we dont need this - set copies to 0 

def oversample(df_lcs, copies=0):
    df_oversample = df_lcs.copy()
    df_oversample['copy_num'] = 0
    for i in range(1, copies+1):
        df_temp = df_lcs.copy()
        df_temp['copy_num'] = i
        df_temp['Mag'] = np.random.normal(df_lcs.Mag, df_lcs.Magerr)
        df_oversample = df_oversample.append(df_temp)
        
    df_oversample = df_oversample.set_index(['copy_num'], append=True)
    return df_oversample

Oversample __supernovas__ light curves

df_sns_0_os = oversample(df_sn_5, 0)
print_num_ids_shape(df_sns_0_os)

del df_sn_5

Oversample __non-supernovas__ light curves

df_nsns_0_os = oversample(df_nsns_5, 0)
print_num_ids_shape(df_nsns_0_os)

del df_nsns_5

Overssample __measured supernova__ light curves

df_msns_0_os = oversample(df_msns_5, 0)
print_num_ids_shape(df_msns_0_os)

del df_msns_5

#### Feature Extraction

def extract_features(df_sns_0_os):
    num_copy_list = df_sns_0_os.index.get_level_values('copy_num').unique()    
    num_copies = len(num_copy_list)

    # Extract IDs list
    unique_ids_list = df_sns_0_os.index.get_level_values('ID').unique()
    num_ids = len(unique_ids_list)

    # Extract bandpass list
    unique_bandpass_list = df_sns_0_os.index.get_level_values('BANDPASS').unique()
    num_bandpass = len(unique_bandpass_list)

    # Create empty feature dict
    feats_dict = extract.feature_dict(30)
    feats_dict['ObsCount'] = []
    feats_dict['Class'] = []

    # Add 'ID', 'copy_num' and 'BANDPASS' index lists
    index_id_list = []
    index_copy_num_list = []
    index_bandpass_list =  []

    num_objects = num_ids*num_copies*num_bandpass
    for num_copy in num_copy_list:
        for i, obj_id in enumerate(unique_ids_list):
            for j, bandpass_id in enumerate(unique_bandpass_list):

                # Print status
                current_object_i = (num_copy+1)*i + j
                if(current_object_i%int(num_objects/10) == 0):
                    print(current_object_i, '/', num_objects)

                # Get current object light curve
                df_object = df_sns_0_os.loc[obj_id,:,bandpass_id,num_copy]
                if(len(df_object)==0): 
                    continue

                # Get features
                obj_feats = extract.features(df_object, feats_dict)

                # Append features
                for k,v in obj_feats.items():
                    feats_dict[k].append(obj_feats[k])

                # Append Indexes
                index_id_list.append(obj_id)
                index_copy_num_list.append(num_copy)
                index_bandpass_list.append(bandpass_id)

                # Append class and obs_count
                assert(len(df_object['class'].unique()) == 1)
                feats_dict['Class'].append(df_object['class'].unique()[0])
                feats_dict['ObsCount'].append(df_object.shape[0])

    # Create feature dataframe
    df_feats = pd.DataFrame(feats_dict).set_index([index_id_list,index_copy_num_list,index_bandpass_list])
    df_feats.index.names = ['ID', 'copy_num','BANDPASS']
    return df_feats

def save_features(df_feats, obj_type,filename):
    outdir = FEATURES_PATH
    filename_raw = '{}.pickle'
    filename = filename_raw.format(filename)
    assert(df_feats.shape[1]==32) # 30 + ['num_obs'+'class']
    df_feats.to_pickle(outdir + filename)

#### Generate Features

def generate_features(df_all, sn, min_obs,output_filename):
    obj_type = 'SN' if sn else 'NSN'
    
    # Generate features based on light curves
    df_feats = extract_features(df_all)
    
    # Save features
    save_features(df_feats, obj_type,output_filename)
    
    # Log Finished
    print('Finished task type={} obs={}'.format(obj_type, min_obs) )
    
    return df_feats

Generate features __transient__ light curves

sns_feats = generate_features(df_sns_0_os, sn=True, min_obs=5,output_filename="SN")

nsns_feats = generate_features(df_nsns_0_os, sn=False, min_obs=5,output_filename="NSN")

msns_feats = generate_features(df_msns_0_os,sn=True,min_obs=5,output_filename="MSN")
