==> 1.1 Group all filters.ipynb <==
import pandas as pd

DATA_PATH = './data/'
LIGHTCURVES_PATH = DATA_PATH + 'lightcurves/'
FEATURES_PATH = DATA_PATH + 'features/'
FEATURES_4D = DATA_PATH + '4dFeatures/'
JUST_RED = DATA_PATH + 'just_red/'

def colapseFilters(df,objType):
    idList = list(df.index.get_level_values('ID').unique())
    filterList= ['i','r','z','g']
    cols = ["ID"]
    original_cols = list(df.columns)
    new_cols = []
    for iFilter in filterList:
        new_cols += [col + "_"+iFilter for col in original_cols]
    cols += new_cols
    alldf = pd.DataFrame(columns = cols)

    for iter_num,iD in enumerate(idList):
        #print status
        if(iter_num%((len(idList)/100)//1)==0 ):
            print("{}% done".format(iter_num/((len(idList)/100)//1)) )
            
#         print("-{}-".format(iD))
        
        item = df.loc[iD,:,:]

        values = []
        for iFilter in filterList:

            #Because of Diegos algorithm, we have that copy_num always has value 0
            #Thus, we can ignore it and get all values of the item using item.loc[0]
            try:
                values += list(item.loc[0].loc[iFilter])

            #An exception means that there is no filter in that object, thus the values
            #are set to 0
            except:
                values += [0.0]*32

        values = [iD] + values

        alldf.loc[len(alldf)]= values
        
    alldf = alldf.drop(['Class_i','Class_r','Class_z','Class_g'], axis=1)
    alldf['Class'] = objType

    return alldf

#must delete class columns and only have 1!!

filenames = ['SN','MSN','NSN']
dfs = {}
for name in filenames:
    dfs[name] =  pd.read_pickle(FEATURES_PATH+name+".pickle")
    


dfs["SN"].shape

allsns = colapseFilters(dfs['SN'],'sn')

allsns.to_pickle(FEATURES_4D + "allSNS.pickle")

allmsn = colapseFilters(dfs['MSN'],'sn')
allmsn.to_pickle(FEATURES_4D + "allMSN.pickle")


allnsn = colapseFilters(dfs['NSN'],'nsn')
allnsn.to_pickle(FEATURES_4D + "allNSN.pickle")

dfs["NSN"].loc[dfs["NSN"].index.get_level_values("ID").unique()[0],:,:]

## Getting only red filters

sn_red = dfs["SN"][dfs["SN"].index.get_level_values('BANDPASS') == 'r']

dfs["SN"].shape

sn_red.shape

sn_red.to_pickle(JUST_RED + 'sn_just_red.pickle')

nsn_red = dfs["NSN"][dfs["NSN"].index.get_level_values('BANDPASS') == 'r']

nsn_red.shape

nsn_red.to_pickle(JUST_RED + 'nsn_just_red.pickle')

msn_red = dfs["MSN"][dfs["MSN"].index.get_level_values('BANDPASS') == 'r']

msn_red.to_pickle(JUST_RED+"msn_just_red.pickle")
