==> 2. Generate Inputs.ipynb <==
%reload_ext autoreload
%autoreload 2

import pandas as pd
import numpy as np
import inputs2

DATA_PATH = './data/'
FEATURES_PATH = DATA_PATH + 'features/'
INPUTS_PATH = DATA_PATH + 'inputs/'

#### Load Transient Features

filename = 'SN.pickle'
indir = FEATURES_PATH; filepath = indir + filename
df_feats_sn = pd.read_pickle(filepath)
df_feats_sn = df_feats_t.sort_index(level=['copy_num', 'ID'])

#### Load Non-Transient Features

filename = 'NSN.pickle'
indir = FEATURES_PATH; filepath = indir + filename
df_feats_nsn = pd.read_pickle(filepath)

df_feats_nsn.columns

inputs2.generate_and_save_all_inputs(df_feats_t, df_feats_nt)
