==> 3.1 Classification 4 filters.ipynb <==
from sklearn.ensemble import RandomForestClassifier
import sklearn.metrics
import pandas as pd
import numpy as np
np.random.seed(0)
from sklearn.model_selection import GridSearchCV, StratifiedKFold  # , cross_val_score
from sklearn.metrics import make_scorer, accuracy_score, precision_score, recall_score, f1_score, precision_recall_fscore_support, confusion_matrix  # , classification_report



DATA_PATH = "./data/"
FEATURES_4D = DATA_PATH + "4dFeatures/"

sn_train = pd.read_pickle(FEATURES_4D + 'allSNS.pickle')

sn_train.shape

sn_test = pd.read_pickle(FEATURES_4D + "allMSN.pickle")

sn_test.shape


allnsn_small = pd.read_pickle(FEATURES_4D+"allNSNsmall.pickle")

allnsn_small.shape

nsns_test = allnsn_small.sample(300)

nsns_train = allnsn_small[~allnsn_small.ID.isin(nsns_test.ID)].sample(5000)

nsns_test.shape

all_train = pd.concat([sn_train,nsns_train])

all_test = pd.concat([sn_test,nsns_test])

all_train.shape

all_train.head()

all_test.head()

features = all_train.columns[1:-1]

y_train = pd.factorize(all_train['Class'])[0]
y_train




clf = RandomForestClassifier(n_estimators=50,random_state=0, class_weight='balanced')

Hay nan values....

all_train = all_train.fillna(0)

clf.fit(all_train[features], y_train)

all_test = all_test.fillna(0)

preds = clf.predict(all_test[features])

y_test = pd.factorize(all_test["Class"])[0]

pd.crosstab(preds, y_test, rownames=['Predicted '], colnames=['Actual '])
