==> 4. Exploration.ipynb <==
%reload_ext autoreload
%autoreload 2

import exploration as explore
import generate_latex

generate_latex.generate_latex_results()

#### Select and Run top task we want to explore

# Select
top_func = exploration.top_eight_c

# Run
best_data, best_model = top_func()
best_data

#### Confusion Matrix

explore.print_confusion_matrix(best_data)

#### Classification Test Scores

explore.print_classification_results(best_data)

#### Feature Importances

best_importances = explore.feature_importance(best_model)
explore.print_feature_importance(best_importances)
